Move test utils into ansiblelint.testing
Moves our few test utils inside a library module in order to make
them available to other consumers, like rules.
