+++ test/TestComparisonToEmptyString.py
@@ -2,8 +2,7 @@ import unittest
 
 from ansiblelint.rules import RulesCollection
 from ansiblelint.rules.ComparisonToEmptyStringRule import ComparisonToEmptyStringRule
-
-from . import RunFromText
+from ansiblelint.testing import RunFromText
 
 SUCCESS_TASKS = '''
 - name: shut down
