+++ test/TestEnvVarsInCommand.py
@@ -2,8 +2,7 @@ import unittest
 
 from ansiblelint.rules import RulesCollection
 from ansiblelint.rules.EnvVarsInCommandRule import EnvVarsInCommandRule
-
-from . import RunFromText
+from ansiblelint.testing import RunFromText
 
 SUCCESS_PLAY_TASKS = '''
 - hosts: localhost
