+++ test/TestMetaChangeFromDefault.py
@@ -2,8 +2,7 @@ import unittest
 
 from ansiblelint.rules import RulesCollection
 from ansiblelint.rules.MetaChangeFromDefaultRule import MetaChangeFromDefaultRule
-
-from . import RunFromText
+from ansiblelint.testing import RunFromText
 
 DEFAULT_GALAXY_INFO = '''
 galaxy_info:
