+++ test/TestUseHandlerRatherThanWhenChanged.py
@@ -4,8 +4,7 @@ from ansiblelint.rules import RulesCollection
 from ansiblelint.rules.UseHandlerRatherThanWhenChangedRule import (
     UseHandlerRatherThanWhenChangedRule,
 )
-
-from . import RunFromText
+from ansiblelint.testing import RunFromText
 
 SUCCESS_TASKS = '''
 - name: print helpful error message
